requirements.txt
pygame-ce
moderngl
numpy